ai_server/requirements.txt
flask>=3.0
torch>=2.0
torchvision
trimesh
Pillow
numpy
google-genai>=1.0
open3d>=0.18
rembg
timm
einops
huggingface-hub
